Relocate `Lighting` tab contents
- Move `Sky exposure`, `Sky light intensity`, and `Apparent sky brightness` to the `Sky & Fog` tab.
- Add an `Apparent sun size` control.
- Add a `Sun` tab.
- Move all sun controls to the `Sun` tab.
- Rename the `Lighting` tab to the `Emitters` tab.
- Add a hint about material emittance to the `Emitters` tab.
- Move ESS controls from the `Advanced` tab to the `Emitters` tab.
- Disable controls when their effects are nullified by the value of another control.
- Relocated cloud controls to a collapsible panel called `Cloud settings`.
- Set the panel to be hidden if `Enable clouds` is disabled.
- Updated some tooltips.
- Increase the wrapping width of the hint in the `Postprocessing` tab.
