+++ chunky/src/res/se/llbit/chunky/ui/render/tabs/EmittersTab.fxml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.Separator?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.control.ScrollPane?>
+<?import se.llbit.chunky.ui.DoubleAdjuster?>
+<?import se.llbit.chunky.ui.IntegerAdjuster?>
+<?import javafx.scene.control.ChoiceBox?>
+<?import javafx.scene.text.Text?>
+
+<fx:root type="javafx.scene.control.ScrollPane" xmlns="http://javafx.com/javafx/8.0.40" xmlns:fx="http://javafx.com/fxml/1">
+  <VBox spacing="10.0">
+    <CheckBox fx:id="enableEmitters" mnemonicParsing="false" text="Enable emitters" />
+    <DoubleAdjuster fx:id="emitterIntensity" maxWidth="1.7976931348623157E308" />
+    <Text strokeType="OUTSIDE" strokeWidth="0.0" wrappingWidth="350" text="For more granular control over material emittance, use the emittance controls in the Materials tab." />
+    <Separator />
+    <HBox alignment="CENTER_LEFT" spacing="10.0">
+      <Label text="Emitter Sampling Strategy:" />
+      <ChoiceBox fx:id="emitterSamplingStrategy" />
+    </HBox>
+    <IntegerAdjuster fx:id="gridSize" />
+    <CheckBox fx:id="preventNormalEmitterWithSampling" mnemonicParsing="false" text="Prevent normal emitter when using emitter sampling" />
+    <padding>
+      <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+    </padding>
+  </VBox>
+</fx:root>
