+++ chunky/src/res/se/llbit/chunky/ui/render/tabs/PostprocessingTab.fxml
@@ -16,7 +16,7 @@
       <Label text="Postprocessing filter:" />
       <ChoiceBox fx:id="postprocessingFilter" prefWidth="150.0" />
     </HBox>
-    <Text strokeType="OUTSIDE" strokeWidth="0.0" wrappingWidth="275"
+    <Text strokeType="OUTSIDE" strokeWidth="0.0" wrappingWidth="350"
       text="Postprocessing affects performance when Render Preview tab is visible. Switching to the Map tab mitigates this." />
     <padding>
       <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
